Split Server header to separate utilities.
