+++ source/httpserver/Server.cpp
@@ -1,4 +1,6 @@
 #include "Server.h"
+#include "ServerUtilities.h"
+#include "Socket.h"
 #include <csignal>
 #include <sstream>
 #include <string>
@@ -6,34 +8,66 @@
 
 namespace {
 
-    std::map<Http::Response::Status, std::string> StockResponse = {
-    { Http::Response::Status::Ok, "HTTP/1.0 200 OK\r\n" },
-    { Http::Response::Status::Created, "HTTP/1.0 201 Created\r\n" },
-    { Http::Response::Status::Accepted, "HTTP/1.0 202 Accepted\r\n" },
-    { Http::Response::Status::NoContent, "HTTP/1.0 204 No Content\r\n" },
-    { Http::Response::Status::MultipleChoices, "HTTP/1.0 300 Multiple Choices\r\n" },
-    { Http::Response::Status::MovedPermanently, "HTTP/1.0 301 Moved Permanently\r\n" },
-    { Http::Response::Status::Found, "HTTP/1.0 302 Found\r\n" },
-    { Http::Response::Status::NotModified, "HTTP/1.0 304 Not Modified\r\n" },
-    { Http::Response::Status::BadRequest, "HTTP/1.0 400 Bad request\r\n" },
-    { Http::Response::Status::Unauthorized, "HTTP/1.0 401 Unauthorized\r\n" },
-    { Http::Response::Status::Forbidden, "HTTP/1.0 403 Forbidden\r\n" },
-    { Http::Response::Status::NotFound, "HTTP/1.0 404 Not Found\r\n" },
-    { Http::Response::Status::RequestTimeout, "HTTP/1.0 408 Request Timeout\r\n" },
-    { Http::Response::Status::InternalServerError, "HTTP/1.0 500 Internal Server Error\r\n" },
-    { Http::Response::Status::NotImplemented, "HTTP/1.0 501 Not Implemented\r\n" },
-    { Http::Response::Status::BadGateway, "HTTP/1.0 502 Bad Gateway\r\n" },
-    { Http::Response::Status::ServiceUnavailable, "HTTP/1.0 503 Service Unavailable\r\n" }
+    std::map<Http::ResponseStatus, std::string> StockResponse = {
+    { Http::ResponseStatus::Ok, "HTTP/1.0 200 OK\r\n" },
+    { Http::ResponseStatus::Created, "HTTP/1.0 201 Created\r\n" },
+    { Http::ResponseStatus::Accepted, "HTTP/1.0 202 Accepted\r\n" },
+    { Http::ResponseStatus::NoContent, "HTTP/1.0 204 No Content\r\n" },
+    { Http::ResponseStatus::MultipleChoices, "HTTP/1.0 300 Multiple Choices\r\n" },
+    { Http::ResponseStatus::MovedPermanently, "HTTP/1.0 301 Moved Permanently\r\n" },
+    { Http::ResponseStatus::Found, "HTTP/1.0 302 Found\r\n" },
+    { Http::ResponseStatus::NotModified, "HTTP/1.0 304 Not Modified\r\n" },
+    { Http::ResponseStatus::BadRequest, "HTTP/1.0 400 Bad request\r\n" },
+    { Http::ResponseStatus::Unauthorized, "HTTP/1.0 401 Unauthorized\r\n" },
+    { Http::ResponseStatus::Forbidden, "HTTP/1.0 403 Forbidden\r\n" },
+    { Http::ResponseStatus::NotFound, "HTTP/1.0 404 Not Found\r\n" },
+    { Http::ResponseStatus::RequestTimeout, "HTTP/1.0 408 Request Timeout\r\n" },
+    { Http::ResponseStatus::InternalServerError, "HTTP/1.0 500 Internal Server Error\r\n" },
+    { Http::ResponseStatus::NotImplemented, "HTTP/1.0 501 Not Implemented\r\n" },
+    { Http::ResponseStatus::BadGateway, "HTTP/1.0 502 Bad Gateway\r\n" },
+    { Http::ResponseStatus::ServiceUnavailable, "HTTP/1.0 503 Service Unavailable\r\n" }
     };
 
-    Tcp::ConstBuffer MakeBuffer(Http::Response::Status status)
+    Tcp::ConstBuffer MakeBuffer(Http::ResponseStatus status)
     {
         return Tcp::MakeBuffer(const_cast<const std::string&>(StockResponse[status]));
     }
 }
 
+namespace Http {
 
-Http::Connection::Connection(Tcp::Socket socket, HandlerStrategy& handler) : socket(std::move(socket)), globalHandler(handler)
+struct Connection::ConnectionPimpl
+{
+    ConnectionPimpl(Tcp::Socket socket) : socket(std::move(socket))
+    {
+    }
+
+    Tcp::Socket socket;
+    RequestParser parser;
+    Request request;
+};
+
+struct Server::ServerPimpl
+{
+    ServerPimpl(ServicePtr service) : 
+        service(std::move(service)),
+        acceptor(*this->service),
+        signals(*this->service)
+    {
+    }
+
+    Server::ServicePtr service;
+    Tcp::Acceptor acceptor;
+    Tcp::SignalSet signals;
+};
+
+}
+
+Http::Connection::Connection(Tcp::Socket socket, HandlerStrategy& handler) : pimpl(new ConnectionPimpl(std::move(socket))), globalHandler(handler)
+{
+}
+
+Http::Connection::~Connection()
 {
 }
 
@@ -44,23 +78,23 @@ void Http::Connection::start()
 
 void Http::Connection::stop()
 {
-    socket.close();
+    pimpl->socket.close();
 }
 
 void Http::Connection::read()
 {
     auto self = shared_from_this();
-    socket.asyncReadSome(Tcp::MakeBuffer(buffer),
+    pimpl->socket.asyncReadSome(Tcp::MakeBuffer(buffer),
         [this, self](int ec, std::size_t bytes)
     {
         if (!ec) // Nie wykryto błędu.
         {
             RequestParser::Result result;
             auto it = buffer.begin();
-            std::tie(result, it) = parser.parse(buffer.begin(), buffer.begin() + bytes, request);
+            std::tie(result, it) = pimpl->parser.parse(buffer.begin(), buffer.begin() + bytes, pimpl->request);
             if (result == RequestParser::Result::Good) // Zapytanie sparsowane poprawnie.
             {
-                if (parser.fill(it, buffer.begin() + bytes, request)) // Zacznij czyta? cia?o.
+                if (pimpl->parser.fill(it, buffer.begin() + bytes, pimpl->request)) // Zacznij czyta? cia?o.
                 {
                     write(); // Jeżeli ciało nie zostało wykryte, przejdź do odpowiedzi.
                 }
@@ -71,9 +105,9 @@ void Http::Connection::read()
             }
             else if (result == RequestParser::Result::Bad) // Zapytanie sparsowane niepoprawnie.
             {
-                globalHandler.respond(socket, Response::Status::BadRequest); // Od razu odpowiedz klientowi.
-                socket.shutdown();
-                socket.close(); // Zamknij gniazdo.
+                globalHandler.respond(pimpl->socket, ResponseStatus::BadRequest); // Od razu odpowiedz klientowi.
+                pimpl->socket.shutdown();
+                pimpl->socket.close(); // Zamknij gniazdo.
                 globalHandler.stop(shared_from_this());
             }
             else
@@ -86,13 +120,13 @@ void Http::Connection::read()
             try
             {
                 if (ec > 1)
-                    globalHandler.respond(socket, Response::Status::RequestTimeout);
+                    globalHandler.respond(pimpl->socket, ResponseStatus::RequestTimeout);
             }
             catch (const Tcp::SendError&)
             {
 
             }
-            socket.shutdown();
+            pimpl->socket.shutdown();
             globalHandler.stop(shared_from_this());
         }
     }
@@ -102,12 +136,12 @@ void Http::Connection::read()
 void Http::Connection::readBody()
 {
     auto self = shared_from_this();
-    socket.asyncReadSome(Tcp::MakeBuffer(buffer),
+    pimpl->socket.asyncReadSome(Tcp::MakeBuffer(buffer),
         [this, self](int ec, std::size_t bytes)
     {
         if (!ec)
         {
-            if (parser.fill(buffer.begin(), buffer.begin() + bytes, request))
+            if (pimpl->parser.fill(buffer.begin(), buffer.begin() + bytes, pimpl->request))
             {
                 write();
             }
@@ -120,13 +154,13 @@ void Http::Connection::readBody()
         {
             try
             {
-                globalHandler.respond(socket, Response::Status::RequestTimeout);
+                globalHandler.respond(pimpl->socket, ResponseStatus::RequestTimeout);
             }
             catch (const Tcp::SendError&)
             {
 
             }
-            socket.shutdown();
+            pimpl->socket.shutdown();
             globalHandler.stop(shared_from_this());
         }
     }
@@ -141,7 +175,7 @@ void Http::Connection::write()
         {
             try
             {
-                socket.write(Tcp::MakeBuffer(handler(request).raw()));
+                pimpl->socket.write(Tcp::MakeBuffer(handler(pimpl->request).raw()));
             }
             catch (const Tcp::SendError&)
             {
@@ -149,7 +183,7 @@ void Http::Connection::write()
             }
         }
     );
-    socket.shutdown();
+    pimpl->socket.shutdown();
     globalHandler.stop(shared_from_this());
 }
 
@@ -163,12 +197,12 @@ void Http::ThreadedHandlerStrategy::handle(ConnectionResponse response)
     {
         response([](const Request&)
         {
-            return Response(Response::Status::InternalServerError, "", "");
+            return Response(ResponseStatus::InternalServerError, "", "");
         });
     }
 }
 
-void Http::ThreadedHandlerStrategy::respond(Tcp::Socket& socket, Response::Status stockResponse)
+void Http::ThreadedHandlerStrategy::respond(Tcp::Socket& socket, ResponseStatus stockResponse)
 {
     socket.write(MakeBuffer(stockResponse));
 }
@@ -184,7 +218,7 @@ void Http::ThreadedHandlerStrategy::stop(ConnectionPtr connection)
     connections.erase(connection);
 }
 
-Http::Response::Response(Status code, const BodyType & content, const MediaType & mediaType) : responseStatus(code), response(content)
+Http::Response::Response(ResponseStatus code, const BodyType & content, const MediaType & mediaType) : responseStatus(code), response(content)
 {
     if (content.length() > 0) // w przypadku braku ciała nie uzupełniaj nagłówków.
     {
@@ -206,7 +240,7 @@ std::string Http::Response::raw() const
     return retval;
 }
 
-Http::Response::Status Http::Response::status() const
+Http::ResponseStatus Http::Response::status() const
 {
     return responseStatus;
 }
@@ -571,7 +605,7 @@ const Http::HeaderContainer& Http::Request::headers() const
     return headerCollection;
 }
 
-const Http::BodyType Http::Request::body() const
+const Http::BodyType& Http::Request::body() const
 {
     return content;
 }
@@ -713,22 +747,20 @@ std::vector<std::string> Http::Uri::segments() const
 }
 
 Http::Server::Server(const std::string& host, const std::string& port, ServicePtr service, StrategyPtr globalHandler) :
-    service(std::move(service)),
-    acceptor(*this->service),
-    signals(*this->service),
+    pimpl(new ServerPimpl(std::move(service))),
     globalHandler(std::move(globalHandler))//(handler)
 {
-    signals.add(SIGINT); // nie wspierane na Windows
-    signals.add(SIGTERM);
+    pimpl->signals.add(SIGINT); // nie wspierane na Windows
+    pimpl->signals.add(SIGTERM);
 #if defined(SIGQUIT)
     signals.add(SIGQUIT);
 #endif // defined(SIGQUIT)
 
-    Tcp::Endpoint endpoint = this->service->getFactory()->resolve(host, port);
-    acceptor.open(endpoint.protocol());
-    acceptor.setOption(Tcp::Option::ReuseAddress(true));
-    acceptor.bind(endpoint.address());
-    acceptor.listen(50);
+    Tcp::Endpoint endpoint = this->pimpl->service->getFactory()->resolve(host, port);
+    pimpl->acceptor.open(endpoint.protocol());
+    pimpl->acceptor.setOption(Tcp::Option::ReuseAddress(true));
+    pimpl->acceptor.bind(endpoint.address());
+    pimpl->acceptor.listen(50);
     accept();
 }
 
@@ -737,14 +769,23 @@ Http::Server::Server(const std::string & host, const std::string & port, Request
 {
 }
 
+Http::Server::Server(const std::string & host, const std::string & port, RequestHandler handler)
+    : Server(host, port, ServicePtr(new Tcp::StreamService()), StrategyPtr(new ThreadedHandlerStrategy(std::move(handler))))
+{
+}
+
+Http::Server::~Server()
+{
+}
+
 int Http::Server::run()
 {
-    return service->run();
+    return pimpl->service->run();
 }
 
 void Http::Server::accept()
 {
-    acceptor.asyncAccept([this](Tcp::Socket socket)
+    pimpl->acceptor.asyncAccept([this](Tcp::Socket socket)
     {
         globalHandler->start(std::make_shared<Connection>(std::move(socket), *globalHandler));
         accept();
